+++ src/utils/types/student_teacher/student.py
@@ -1,12 +1,14 @@
-from .people import Person
+from .person import Person
 
 from pydantic import BaseModel
 from typing import List, Optional
+
 
 class Student(BaseModel):
     """
     Student model
     """
+
     id: str
     people: Person
     std_id: str
