+++ src/utils/types/student_teacher/teacher.py
@@ -1,14 +1,16 @@
-from .people import Person
+from .person import Person
 
 from pydantic import BaseModel
 from typing import List, Optional
 
 from subject.subject_group import SubjectGroup
 
+
 class Teacher(BaseModel):
     """
     Teacher model
     """
+
     id: str
     people: Person
     teacher_id: str
